requirements.txt
Package             Version
------------------- ---------
anyio               4.6.0
APScheduler         3.10.4
beautifulsoup4      4.12.3
certifi             2024.8.30
cffi                1.17.1
charset-normalizer  3.3.2
cryptography        43.0.1
exceptiongroup      1.2.2
greenlet            3.1.1
h11                 0.14.0
httpcore            1.0.6
httpx               0.27.2
idna                3.10
pillow              10.4.0
pip                 22.0.2
psycopg2-binary     2.9.9
pycparser           2.22
python-dotenv       1.0.1
python-telegram-bot 21.6
pytz                2024.2
requests            2.32.3
setuptools          59.6.0
six                 1.16.0
sniffio             1.3.1
soupsieve           2.6
SQLAlchemy          2.0.35
typing_extensions   4.12.2
tzlocal             5.2
urllib3             2.2.3
wheel               0.37.1
